refactor: 예외 처리 한 곳에서 하도록 변경
